% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/semieffsdr.R
\name{effsdr}
\alias{effsdr}
\title{Estimate the dimension reduction direction via semiparametric efficient estimator}
\usage{
effsdr(X, Y, ndr, beta_init)
}
\arguments{
\item{X}{predictor matrix}

\item{Y}{Response matrix}

\item{ndr}{Dimension of directions selected}

\item{beta_init}{Initial value of beta}
}
\description{
Estimate the dimension reduction direction via semiparametric efficient estimator
}
\examples{
set.seed(123)
n=500
X=cbind(rnorm(n,sd=10),rnorm(n,sd=10),rnorm(n,sd=1),rnorm(n,sd=1))
Y=I(X[,1]+X[,2]>0)+rnorm(n,sd=0.5)
beta_init= orthoDr::initB(X, as.matrix(Y), ndr=1, bw = 1, method = "sir", ncore=0)
effsdr(X,Y,1,beta_init)
}
\references{
Ma, Y., & Zhu, L. (2012). A semiparametric approach to dimension reduction. Journal of the American Statistical Association, 107(497), 168-179.

Ma, Y., & Zhu, L. (2013). Efficient estimation in sufficient dimension reduction. Annals of statistics, 41(1), 250.
}
