% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GMMboot.R
\name{GMMboot}
\alias{GMMboot}
\title{Compute the mean of squared vector correlation among enveloped central space of bootstrap samples estimated by GMM method}
\usage{
GMMboot(X_data, Y_data, u, B)
}
\arguments{
\item{X_data}{predictor matrix}

\item{Y_data}{Response matrix}

\item{u}{Dimension of envelope selected}

\item{B}{Number of bootstraps}
}
\description{
Compute the mean of squared vector correlation among enveloped central space of bootstrap samples estimated by GMM method
}
