% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GMMboot.R
\name{GMMdim}
\alias{GMMdim}
\title{Estimate the dimension of Enveloped central space via GMM estimator}
\usage{
GMMdim(X, Y, B)
}
\arguments{
\item{X}{predictor matrix}

\item{Y}{Response matrix}

\item{B}{Number of bootstraps}
}
\description{
Estimate the dimension of Enveloped central space via GMM estimator
}
