% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Globaleff.R
\name{localeff}
\alias{localeff}
\title{Estimate enveloped central space via local efficient estimator}
\usage{
localeff(X, Y, u, gamma_init = "notgiven", beta)
}
\arguments{
\item{X}{predictor matrix}

\item{Y}{Response matrix}

\item{u}{Dimension of envelope selected}

\item{gamma_init}{Initial value of gamma, defaultly obtaining from GMM method}

\item{beta}{Model specification factor}
}
\description{
Estimate enveloped central space via local efficient estimator
}
\examples{
set.seed(300)
Gamma0 = diag(4)[,c(1,4)]
Gamma = diag(4)[,c(2,3)]
Omega = matrix(1, nrow = 2, ncol = 2) + diag(2, 2)
Omega0 = diag(c(2,2))
SigmaX = Gamma \%*\% Omega \%*\% t(Gamma) + Gamma0 \%*\% Omega0 \%*\% t(Gamma0)
X=mvrnorm(300, mu = rep(0, 4), Sigma = SigmaX, tol=1e-6)
Y=numeric()
for(i in 1:300)
Y[i]=sample(c(0,X[i,3]/2,X[i,3]),1)+rnorm(1)
result=localeff(X,Y,2,beta=c(1,1,1,1))
}
