% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nGMM2.R
\name{nGMMenv}
\alias{nGMMenv}
\title{Estimate enveloped central space via GMM estimator}
\usage{
nGMMenv(X, Y, u)
}
\arguments{
\item{X}{predictor matrix}

\item{Y}{Response matrix}

\item{u}{Dimension of envelope selected}
}
\description{
Estimate enveloped central space via GMM estimator
}
\examples{
#generate  data
set.seed(1)
Gamma = diag(3)[,1:2]
Gamma0 = diag(3)[,3]
Omega = matrix(1, nrow = 2, ncol = 2) + diag(2, 2)
Omega0 = diag(0.1, 1)
SigmaX = Gamma \%*\% Omega \%*\% t(Gamma) + Gamma0 \%*\% Omega0 \%*\% t(Gamma0)
X=mvrnorm(100, mu = rep(0, 3), Sigma = SigmaX, tol=1e-6)
Y=X[,1]+rnorm(100)
Get envelope space
result=nGMMenv(X,Y,2)
Gammaest=result$Gamma1
Gammaest
}
