% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nGMM2.R
\name{nGMMenv3}
\alias{nGMMenv3}
\title{Estimate enveloped central space via GMM estimator, tangent scaling parameter}
\usage{
nGMMenv3(X, Y, u)
}
\arguments{
\item{X}{predictor matrix}

\item{Y}{Response matrix}

\item{u}{Dimension of envelope selected}
}
\description{
Estimate enveloped central space via GMM estimator, tangent scaling parameter
}
\examples{
#generate  data
set.seed(3)
Gamma = diag(5)[,c(1,4)]
Gamma0 = diag(5)[,c(2,3,5)]
Omega = matrix(1, nrow = 2, ncol = 2) + diag(2, 2)
Omega0 = diag(c(0.2,0.2,2))
SigmaX = Gamma \%*\% Omega \%*\% t(Gamma) + Gamma0 \%*\% Omega0 \%*\% t(Gamma0)
X=mvrnorm(500, mu = rep(0, 5), Sigma = SigmaX, tol=1e-6)
Y=exp(X[,1])+8*I(X[,3]>0)+rnorm(500)
#Get Envelope Space
result=nGMMenv3(X,Y,3)
Gammaest=result$Gamma1
Gammaest
}
