% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Globaleff.R
\name{P}
\alias{P}
\title{Compute the projection matrix onto the column space of the design matrix A}
\usage{
P(A)
}
\arguments{
\item{A}{the design matrix A}
}
\description{
Compute the projection matrix onto the column space of the design matrix A
}
