% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Globaleff.R
\name{Q}
\alias{Q}
\title{Compute the projection matrix onto the orthogonal complement of the column space of the design matrix A}
\usage{
Q(A)
}
\arguments{
\item{A}{the design matrix A}
}
\description{
Compute the projection matrix onto the orthogonal complement of the column space of the design matrix A
}
