% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/semisave.R
\name{semisave}
\alias{semisave}
\title{Estimate the dimension reduction direction via semi-SAVE estimator}
\usage{
semisave(X, Y, ndr, beta_init)
}
\arguments{
\item{X}{predictor matrix}

\item{Y}{Response matrix}

\item{beta_init}{Initial value of beta}

\item{u}{Dimension of directions selected}
}
\description{
Estimate the dimension reduction direction via semi-SAVE estimator
}
